ModalError: include 'cause' + formatting changes
## History
'cause' is the usually the object that we add to the `Error` so that it gets transmitted to the loggers. We hid it previously from the user since it could be very verbose.

## Change
Considering the error message could be terse at times, and loggers can get blocked, show 'cause' so that users can at least copy/paste.