Publish v4: auto v3-to-v4 transition
## Issue
If there is an incomplete v3 upload, and the page is refreshed to v4, the upload entry is no longer resumable.

## Change
Since we still have the params, re-upload from scratch to v4 behind the scenes.

If the SDK call has already been made, then don't allow to retry (just 'Remove')