Fix inverted data flow way of setting 'edited claim channel'
## Symptom
If you edit a claim with restrictions, the selected restrictions is reset due to the active channel taking several renders to resolve to the right value. For the user, it would be nice that the previous value is pre-selected.

## Root
The 'channelId' of the form was being updated through the weird `channelToSet` mechanism of the ChannelSelector. This will definitely be several renders late, plus data flow is in the opposite direction

## Change
Just like all the recent changes, `doPrepareEdit` is the place to set all the states right before we start rendering the form.