Band-aid: debounce Post Editor (#2908)
## Ticket
2907 - Post Editor laggy with huge articles

## Root
The classic problem with "controlled input" -- each character-change causes the component to render itself.

## Change
Work in "uncontrolled input" mode (which simpleMDE already does) + debounce the redux write.

This is just a band-aid, as there will still be stutter every now and then when the debounced function finally executes during idle (and user coincidently types again).

But at least the user can type now.

## Notes for full fix
The crux of the problem is the dependence on the `fileText` state. Try this:

1. There are 4 effects that fetches the post's data (if in Editing mode). Move this into a single redux action and store the outcome into `publish.postOriginalText`, and copy that into `publish.fileText` in `doPrepareEdit`.
2. In the component, use `publish.postOriginalText` instead of `publish.fileText` as the starting point.
3. Continue to use the debounced function that updates `publish.fileText`.

That way, the component won't be re-rendering on every character typed.