Do not include previews when detecting name-clashes
## Issue
When publishing with preview=on, the name collision warning briefly appears
https://user-images.githubusercontent.com/34790748/260001187-84c5f604-ee2f-4757-88d6-412a5569279c.webm

## Why
When a publish is pending via tus, we prevent the next upload form from re-using any of the pending claim names. As we are replacing v1 with v4, every upload (including previews), is a tus upload now, so it's trying to warning against itself, lol.

## Fix
Just don't include previews in the pending name list.