Publish: attempt to handle replay name-clash failures (#2932)
## Issue
> Did get a complaint from a user who was unable to edit his livestream placeholder to a replay. It would tell him he has a duplicate claim, so it was sending a stream create vs an edit. Trying to repro how they are getting into that scenario, works normally for us otherwise.

## Possible Root
Wasn't able to replicate, but if the issue is about `stream_create` vs. `stream_update`, then that narrows it down to a missing `claim_id` when creating the payload.

The use of `selectMyClaimForUri` to get the claim-being-edited has always been sketchy, since that depends network fetches and a ton of logic. Most likely it returned undefined here.

## Change
With recent changes, we are now storing `claimToEdit` in Redux and persisting it over F5. That object is guaranteed to be populated first before entering Edit stage, so derivation isn't necessary anymore.