Tweak daemonCallWithResult hook for extensibility
(improves 4267c1cc)

No functional change, just refactoring the third parameter to be a general-purpose hook rather than one that is hardcoded for a single-purpose.

The impetus for this is so that we can chain additional hooks cleanly, which we'll need for an upcoming PR to limit `claim_search::not_channel_ids` to 2k.