claim_search: handle blocklist limit
## Issue
> WARNING hub.herald.session:1285: Got an invalid query from ip, for not_channel_ids with more than 2048 elements.

This is most likely coming from a huge user muted/blocked-list.

Homepage data is another possibility, but we don't have anything that large today.

## Change
Limit `not_channel_ids` to 2047 (wasn't sure if it's inclusive of 2048).

## Tradeoff
Some of the blocked list will slip through, but at least the call doesn't fail. A more elaborate change would be required for a full fix, probably at the SDK side?