Hide comments from deleted channels (#2940)
* Remove unused prop

* Hide comments from deleted channels

## Ticket
2934 comments for deleted channel still shown

## Old behavior
- `UriIndicator` will display the deleted channel name as "[Removed]".
- Comment text remains displayed.

## Root cause
2554 (1f60d017) removed `UriIndicator` and tries to display the best channel name format based on `comment` object data, which is populated even if the channel was deleted.

## Change
I think there was a reason why we decided to show the comment instead of hiding it, but I can't recall. Maybe it's so that the Total shown tallies, or some pagination issue.

Anyway, just hide it for now as suggested in the ticket. If an issue is found, we can try to restore the old behavior.