LBC Purchase Modal: repeat fund check instead of failing silently
## Issue
The purchase button invokes the modal, and the modal just fails silently when there's insufficient funds.

## Change
I believe we used to show the button differently when the funds are insufficient (grayed out?).

Given the purchase button + claim overlay refactoring, it's probably easier now to just let the user click the button, and inform in the modal if there's insufficient funds. The file page already have the same "insufficient fund" warning.