Serve missing internal files as 404 instead of soft-404 page
## Issue
Requesting for an invalid js file like `https://odysee.com/public/some-non-existent-file.js` yields HTML, which is usually incorrectly redirect to an actual claim, or a soft-404 page.

It's not only showing incorrectly data, but it is being cached. This is one of the potential causes for the problem of users being stuck with bad cache.

## Approach
- In the final catch-all stage, try to detect if it's a file request, and return a 404 instead.
- Do no cache 404s. Same as https://community.cloudflare.com/t/404s-and-caching/337151. The file could eventually be corrected, but users remain stuck with 404.

## Concerns
- Is the detection too loose?