Defer React script until DOM is ready
## Issue
[LBRY-DESKTOP-WEB-PRD Target container is not a DOM element](https://sentry.lbry.tech/organizations/odysee/issues/28517/)

## Reproduce
Make a bookmark of the site, and keep clicking it with variations in delay

## Root
React looks for a div with a specific ID, so the div must exist first.

## Change
There are various suggestions, from putting the script tag at the bottom-most of <body>, to listening to `DOMContentLoaded`.

The easiest seems to be the `defer` tag -- it's the same as `async` (start fetch in parallel), but only execute it after the DOM is ready.

## Caveat
On Chrome, clicking on a bookmark will utilize the cached version of `index.html` (maybe this is what some users are doing when they say they are "reloading").

The fix will only take effect on F5 (or clicking the button, or clicking our Reload nag).