Add ability for methods to skip the artificial timeout (#3007)
In 9b44b7eb, we added an artificial fetch timeout that's shorter than the browser's default so that (1) user don't have to wait too long (2) we get to differentiate the error as being a timeout (fetch errors all use the same string).

This commit provides the ability to specify list of methods that will bypass that mechanism (i.e. use browser's default timeout length).