Enhances embed functionality for channels/playlists
Extends the embed component to support channels, playlists, and markdown content.

For channel embeds, it now displays a grid of the latest uploads.
Playlist embeds showcase the cover and a playable list.
Adds support for rendering markdown content within embeds, fetching the content and displaying it.

Also configures the homepage embed to point Mini App action to the homepage embed.